Agregada CBV para registrar venta de agua
- El campo 'capacity' del modelo refectory (estaba sin uso) se usará para fijar el precio de venta de agua por comedor.

+++ apps/main/refectories/models.py
@@ -14,8 +14,8 @@ class Refectory(TimeStampedModel):
         blank=True, 
         null=True
         )
-
-    capacity = models.IntegerField(
+    #este campo ahora sera el precio del agua fijado por cada comedor (admin)
+    capacity = models.FloatField(
         blank=True, 
         null=True
         )
